requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
blinker==1.6.2
branca==0.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2022.1
cryptography==2.8
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.1.5
django-cors-headers==3.13.0
django-crispy-forms==2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
ffmpeg-python==0.2.0
Flask==2.3.2
Flask-Cors==3.0.10
folium==0.14.0
future==0.18.3
GDAL==3.0.4
geographiclib==2.0
geoip2==2.9.0
geojson==3.0.1
geopy==2.3.0
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
maxminddb==2.3.0
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==2.0.1
psycopg2==2.9.5
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing-extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.3.4
zipp==3.15.0
